@@ -1,26 +1,39 @@
 {
+  "type": "module",
   "name": "mern-tracker",
   "version": "0.1.0",
   "private": true,
   "dependencies": {
+    "@material-ui/core": "^4.11.4",
     "axios": "^0.21.1",
-    "bootstrap": "^4.3.1",
+    "babel-preset-env": "^1.7.0",
+    "bootstrap": "^4.6.0",
     "cors": "^2.8.5",
     "dotenv": "^10.0.0",
     "express": "^4.17.1",
+    "mongodb": "^3.6.9",
     "mongoose": "^5.12.11",
     "nodemon": "^2.0.7",
-    "react": "^16.8.6",
-    "react-datepicker": "^2.5.0",
-    "react-dom": "^16.8.6",
+    "postcss-cli": "^8.3.1",
+    "react": "^16.14.0",
+    "react-datepicker": "^2.16.0",
+    "react-dom": "^16.14.0",
+    "react-google-login": "^5.2.2",
+    "react-hook-form": "^7.6.9",
+    "react-icons": "^4.2.0",
+    "react-modal": "^3.13.1",
+    "react-router": "^5.2.0",
     "react-router-dom": "^5.2.0",
-    "react-scripts": "3.0.1"
+    "react-scripts": "3.0.1",
+    "tailwind": "^4.0.0"
   },
   "scripts": {
     "start": "react-scripts start",
     "build": "react-scripts build",
     "test": "react-scripts test",
-    "eject": "react-scripts eject"
+    "eject": "react-scripts eject",
+    "build:css": "postcss src/style/tailwind.css -o src/style/main.css",
+    "build:watch": "postcss src/style/tailwind.css -o src/style/main.css --watch"
   },
   "eslintConfig": {
     "extends": "react-app"
@@ -38,7 +51,7 @@
     ]
   },
   "description": "Code for MERN tutorial.",
-  "main": "index.js",
+  "main": "server.js",
   "repository": {
     "type": "git",
     "url": "git+https://github.com/beaucarnes/mern-exercise-tracker-mongodb.git"
@@ -49,5 +62,10 @@
   "bugs": {
     "url": "https://github.com/beaucarnes/mern-exercise-tracker-mongodb/issues"
   },
-  "homepage": "https://github.com/beaucarnes/mern-exercise-tracker-mongodb#readme"
+  "homepage": "https://github.com/beaucarnes/mern-exercise-tracker-mongodb#readme",
+  "devDependencies": {
+    "autoprefixer": "^9.8.6",
+    "postcss": "^7.0.35",
+    "tailwindcss": "npm:@tailwindcss/postcss7-compat@^2.1.2"
+  }
 }
