@@ -1,22 +1,21 @@
-const mongoose = require('mongoose');
+import mongoose from 'mongoose';
 
 const Schema = mongoose.Schema;
 
 const userSchema = new Schema({
-  username: {
+  userName: {
     type: String,
     required: true,
-    unique: true,
     trim: true,
     minlength: 3
   },
-  emailid: {
+  emailId: {
     type: String,
     required: true,
     unique: true,
     trim: true,
   },
-  studid:{
+  studCode:{
     type: String,
     required: true,
     unique: true,
@@ -29,4 +28,4 @@ const userSchema = new Schema({
 
 const User = mongoose.model('User', userSchema);
 
-module.exports = User;
+export default User;
