@@ -1,6 +1,7 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 import App from './App';
-import axios from 'axios'
+
+import './components/style/main.css'
 
 ReactDOM.render(<App />, document.getElementById('root'));
